requirements.txt
discord.py
RPi.GPIO
paho-mqtt